[Feature] Funcionalidade de delete concluída com sucesso

+++ assesment_kreative_user_control/app/Http/Controllers/UserController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
+use Exception;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -26,6 +28,11 @@ class UserController extends Controller
 
     public function deleteUser($id)
     {
+        if($id == Auth::id())
+        {
+            throw new Execption();
+            return "Não podes deletar seu próprio usuário logado";
+        }
         return User::destroy($id);
     }
 }
